Start fixing errors by moving things to use the main info store
