+++ inst/editor/src/components/Tabs/Tabset/TabDropDetector.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { DropWatcherPanel } from "../../../DragAndDropHelpers/DropWatcherPanel";
-import { invalidNavPanelContents } from "../../../ui-node-definitions/Bslib/NavPanel";
+import { invalidNavPanelContents } from "../../../ui-node-definitions/Bslib/nav_panel";
 import type { NodePath } from "../../../ui-node-definitions/NodePath";
 import type { ShinyUiNode } from "../../../ui-node-definitions/ShinyUiNode";
