--- inst/editor/src/ui-node-definitions/Bslib/index.tsx
@@ -1,3 +1,0 @@
-export { bslibCardInfo } from "./BslibCard";
-export { bslibCardBodyInfo } from "./card_body";
-export { bslibCardHeaderInfo } from "./BslibCardHeader";
