--- inst/editor/src/ui-node-definitions/gridlayout/index.ts
@@ -1,5 +1,0 @@
-export { gridlayoutCardInfo } from "./GridlayoutCard";
-export { gridlayoutGridCardPlotInfo } from "./GridlayoutGridCardPlot";
-export { gridlayoutTextPanelInfo } from "./GridlayoutCardText/GridlayoutCardText";
-export { gridlayoutGridContainerInfo } from "./GridlayoutGridContainer";
-export { gridlayoutGridPageInfo } from "./GridlayoutGridPage";
